create Job interface and only expose that not implementation
